Add manifolds for nonlinear spaces in the games Minecraft and Terraria.
In the Minecraft one, it generates a horizontal 'slice' of a Minecraft world, where each node is connected to orthogonal and diagonal neighbours like a king, but each 8\*8 chunk of nodes is connected to a nether node, and nether nodes are connected to each other by king moves covering 8 blocks. The scale of Minecraft blocks to manifold nodes is that at which travelling one unit is the same 'difficulty' as moving to your corresponding position in the nether. Space curves (I think the original idea for the nether was that because it's beneath the overworld, the radius of the negligibly-wide band of the spherical Earth it inhabits is 1/8th of that of the overworld, and the Earth would be canonically so large that the curvature of the area you inhabit would be imperceptible, but it doesn't explain the absence of gravity disparity). A manifold for a square with side-scrolling would be interesting, as well, and it could find ones for Möbius and Klein surfaces.

The Terraria one is a rectangle with the Terraria world's aspect ratio, connected by king moves (this time more accurately than the Minecraft one, because in Terraria x and y speed are independent of each other), you can add one-way connections (that only pull the origin node towards the destination, not bidirectionally) corresponding with teleportation items (the magic mirror, which allows you to move from any node to the origin (and creates s-shaped manifolds), and the conch shell, which lets you move to the opposite shore (and creates more complicatedly shaped manifolds, especially when paired with the magic mirror)).

Both new simulations are in 3D space and, to compensate for 2D initial configurations never protruding into the third dimension on their own, a small initial disparity exists for a random node in both, wait a few seconds and it will accumulate until the entire thing becomes 3D.

I know this commit isn't chess-related but I didn't want to duplicate the program without letting both applications benefit from improvements to the physics (like the 3D mode that prompted the game ones and the future distance-converging mode).