Add quaternion rotation and quotient space scrolling modes
Use the p key for left-right scrolling, o for Möbius of which, i for up-down scrolling, u for Möbius of which, and y to make up instead connect to right (and vice versa) and down connect to left. Allows all Euclidean quotient spaces that don't allow violation of conservation of energy and some that do.
The quaternion rotation allows dragging with the mouse or using the arrow keys when in 3D mode, with the directions being from the current reference frame instead of the origin, much more fun.