Generalise quaternion precomputed rendering index to connections, fix depth being the wrong way around
Precomputed index reduces operations necessary, will speed it up (especially for smaller boards where it (asymptotically proportional to the number of nodes) exceeds n-body repulsion computation (proportional to that squared)). The depth has no 'right' way around, because it's orthographic, but now the side that moves in the same direction as the dragging mouse pointer is in front, seems easier to understand to me.