Replace gravity equations with better ones with component divided by distance cubed instead of trigonometry, new stick-and-ball rendering method
Connections are rendered in front of nodes depending on screen Z, and their origin is a point on the nodes' surfaces instead of centres (seems flickery at the moment).

https://www.wolframalpha.com/input/?i=y%2F%28%28x%5E2+%2B+y%5E2%29%5E%283%2F2%29%29
The solution it gives to this over real numbers but without the sign function in the denominator (think about it for a moment).
Also more concise, I think.