Add spherical linear interpolation (slerp) to the quaternion mode
Because multiplication by fractional degrees works in arcs, not linearly. You can also run the old and broken mode instead, in the render function's connection part, if you want. (Looks more complicated because it orders both nodes and connections in the same list from back to front for rendering, but is actually straightforward.)