Add Warnsdorf heuristic mode
For solving knight's tour, press W to use.

I know the majority of readers of descriptions of commits to a program for simulating chess and graph theory already know about it, but the knight's tour is a puzzle like Ben Finegold's queen-knight one, except instead of moving to each square in order with intermediate moves while avoiding the queen and the squares it attacks, you move to each square in any order while avoiding the squares you've already been to. This is modelled in the program by removing nodes travelled to and simulating the physics in between (the manifold reshaping itself each second correspondingly). Towards the end, when there are fewer nodes, it can form interesting shapes like cubes or pairs of cubes joined together (which I got once when beginning in 16*16) or octahedra (when simulating for pieces without parity-inversion like rooks).

The Warnsdorf heuristic is to look at the nodes you can move to in the next move and count how many moves each of them will have subsequently, and move to the one with the fewest (or a random one if there's a tie). It seems to always solve 8*8 boards (though a paper said its solving-rate reduces and becomes near-0% on 50*50 boards), I will modify it to do an exhaustive search tree of Warnsdorf paths next (of which there aren't very many, it's a poor tour enumeration method compared to the ant colony optimisation one (which I will add also)).

Also, fix node deletion (previously deleted wrong connections due to using the node's indice as both the comparator for the endpoints and the connection indice, not sure how it got away with it for so long, looks very silly now), and fix connection origin indexing (would previously not handle nodes that weren't origins of any connection (it would count the next one's indice as that of the one after it), function now does it correctly by substituting an if statement of whether origin[n]>origin[n-1] for a repeat element[n]-element[n-1]).