Add tablebase vision
A Python program that (less efficiently) solves the superset of the regular chess piece vision program's default mode. Instead of each node representing a possible piece position, it represents that of multiple pieces (not only their permutation but also whose turn it is to move), it computes the attacked squares in each position and thus legal moves, then creates a list for each position of the indexes of the position that each move changes it to. Currently only kings and empty squares implemented and running it displays the tablebase for them on a 3*3 board (you would think it's trivial but there are 64 unique positions ((4 positions where the white king is on a side square (with 3 possible black king positions) and 4 where it's on an edge square (with 5))*2 colours to have the turn) and some one-way state transitions that can cause oscillation in the minimal energy state), but I will add more (and such things as manual dragging and 3D rendering (like the original Scratch version)).