(Adds itertools as dependency) Make state transition graphs work with more pieces than only two kings, finish knight movegen, optimise physics function (and fix slice notation bug where indices were off by one (causing graph embeddings to be extremely suboptimal in previous 3 commits))
itertools is now a dependency because I didn't want to reimplement its combinations and permutations functions myself (though I may in the future), all permutations of each combination of non-king pieces and empty squares fill the gaps that each of the invariant legal king positions create.

(Optimisation done because exponentiating by 1.5 is faster than by 3 then square-rooting with the math function, and di didn't need to be an indice in air resistance and velocity enaction loops.)
Running it now creates a state transition diagram for the 512 possible positions of a king and knight vs. king on 3*3 (in which checkmate is impossible but it converges to the shape of a porcupine or pineapple due to stalemated positions (where the black king is in a corner and the white is on an edge with its knight on a corner adjacent to it) not having any moves acting as springs to pull them back in), I think if you run it for long enough you'll see a 64-node graph escape from the main one (which is the same as the one without a knight, but is where the knight is on the centre square and can't move (like the meme format) and also can't be taken). If you're suspicious of the number being a power of two, there are 64 positions with no knight, 64*6 where the knight is useful and 64 where it's useful, 64*8 in total.

Next, I will add zooming (perhaps not automatic but using screen size better than now (perhaps fitting the minimal circumscribing sphere to the square screen area for determining the centre instead of taking the mean) and zooming based on maximum spatial difference (not screen difference because then it would zoom in and out as you rotate)), iterative pieces (ie. queens/rooks/bishops/nightriders that require the board state to be inputted to them also) and checkmate regression (so it will prune the graph), and thus colour-coding of graph nodes based on who's winning (with a small difference in shade based on whose turn it is) and the ability to play against you (where you input PGN moves) in the command line.