Prune illegal positions (ie. where it's white's turn but black's in check (very suspicious)), index positions by winningness and how many moves they are to mate
Also change the state it generates on startup (again) from 3*3 KNvK to 3*3 KNNvK because it has checkmates (I think even mate-in-8-ply positions but they may be stalemate-in-8 ones). The tablebase (when I implement chess-playing with PGN in the terminal (may make GUI later)) will currently attempt to stalemate as quickly as possible in drawing positions, but I could also make it attempt to swindle (moving to positions where you can't force insufficient material but the proportion of moves you could make after which it mates you is maximised).
Also, I will change enumerate()'s to zip()'s due to them being ~10-to-20% faster.