Prune eightfold symmetry for non-king pieces when kings are on same line of symmetry, fix node occlusion in state transition diagram
More information on title:
Eightfold symmetry previously applied only to put the white king into the bottom-right diagonal half of the bottom-left quadrant of the board (and apply symmetry to the black one to put it on the left/bottom-right side of each line of symmetry it occupies), now other pieces' positions are used (the first disparity in the lists of squares on either side of the line) if both kings are on the same such line. The effects of this are slightly greater on odd board widths due to orthogonal symmetries falling on squares, for KNNNvK in 3*3 there are 4096 possible states, of which 3136 are legal, but with kingwise eightfold symmetry (in generation) there are only 640, of which 500 are legal, of which 482 are unique with piecewise symmetry. Currently the board flips and rotates when playing against God in symmetry mode, will be fixed in the next commit.
Occlusion previously didn't work because it was attempting to sort by the real part of the product of the position vector by the camera angle quaternion (which is always zero) when it should have been using the unused imaginary part (which is the actual nonzero one perpendicular to the screen), by changing the slice notation I accidentally shifted the screen ones, so I had to shift the rotateByScreen input parameters until it worked, now you begin looking at the initial line arrangement from one end and can see it expanding in both perpendicular dimensions instead of only one, which I like more so will keep.

Internal/minor changes not in title:
Make combinations a set (because it's O(n) and more concise) and pieces tuples (they aren't muted currently, the cost of replacing the entire list (containing the colour) instead of changing the first element if promotion is added would be negligible compared to tuples' increased efficiency for all other purposes).
Add kingPositions (to reduce index() usage), remove fixedKingPermutations loop and generate all lists with comprehensions or += instead of repeated .append()'s.
No longer report the number of wins in even numbers of moves and losses in odd numbers (which are always 0).
Use floor division operator everywhere it's applicable.