Add GUI for playing with God
I haven't added Cburnett pieces so currently kings are squares and knights are circles, but moves are highlighted based on whether they're optimal continuations (to win or draw in the fewest moves or lose in the most), sometimes (when you're being checkmated) moving your king away from a knight is green but capturing it is red, because you can't be checkmated either way and drawing can be achieved in fewer moves (assuming cooperation from the opponent (not attempting to maximise the proportion of continuations that lead to winning or accounting for the opponent doing that or anything (drawing optimality isn't important because you draw either way and God doesn't blunder))).
Also, regardless of whether or not you use the GUI mode, it now resets the state to a new random one upon stale/checkmate instead of returning an error if it's God's turn or awaiting an input then returning an error if it's yours. It may be confusing due to God moving and the state resetting instantly and due to eightfold symmetry still making it appear to flip (will be fixed in subsequent commit (sorry)).