Add bishops/queens, optimise regression from losing to winning positions (approximately double speed), fix flipping and crashing
The rook/bishop movegen is currently the same code for both pieces (with conditions switched by repeated list lookups), which slows it down slightly compared to separate functions.
Regression of losing positions doesn't require exhaustive search of each node's descendents, only of each newly found one's parents (to find whether it's not already winning (in which case the existence of the newly-proven losing descendant makes it winning, irrespective of other descendants) or unknown (in which case either a winning or stalemating-by-cooperation move (the two generated on the alternate cycles) is preferred)).
The onscreen moves overlaid are still wrong (I think the eightfold antiflipping accounts for the opponent king as the basis one also on its turn (the king to move is always colour 0, the internal state doesn't XOR by turn) but it may also be because one function is applied in the wrong order), but now clicking works in correspondence with the board, and the turnwise reduction wasn't accounted for previously (ie. referring to currentState[1] for the board), causing crashes, now fixed.
Also, piece rendering has been cleaned up, kings are now octagons instead of squares (which now represent rooks), bishops are squares rotated 45º, queens are octagons rotated 45º, and knights are still circles, pawns will be oriented triangles, I don't know about nightriders (which will be possible through generalising the rook movegen also), perhaps hyperbolas.
Sorry for unimaginative and sparse commits recently.