Optimise state generation by adding black-to-move states and pruning illegalities in-place, make axialDisparity and positionReflect no longer recursive, add compoundReflect (replacing symmetry's mode 3),
State generation no longer needs to record turnwise symmetries of each combination for the black-to-move states (which, in the previous commit, were themselves pruned in-place and computed before but appended only after the white-to-move ones (from which they were derived) were pruned) but export it for curiosity purposes, though stateKingLinesOfSymmetry is no longer stored for all, and because attacks and checks are generated in-place (the only way to prune by them in-place, and also to avoid having to compute them twice for black-to-move check determination), no longer stateKingPositions (which wasn't used for anything other than check detection either). However, turns are now exported from it instead of calculated by list lengths, but this allows for the identification of states in symmetrical combinations with the invention of a new boolean value Tralse, represented in the state transition diagram's default (turn to move) mode with a new colour (the average of light and dark squares).
Three functions are now comprised of binary trees of optimised equations for each of the eight combinations of axial flippings instead of applying each recursively (for instance, positionReflect is about 5 times faster in cProfile). My main motivation for this was making symmetry() work by flipping its directions of comparisons instead of the board inputted, so that it only needs to be flipped if it will be outputted, but also because doing it was fun. However, the old versions (now commented out) now use the reduce() function, since I became aware of it.

When I made the original ones (that are the "pythonic way" as per 'in all other cases it's better to write out the accumulation loop explicitly.') I felt as though there was a more elegant special-case way I was missing. I disagree with Guido van Russom on reduce(), it belongs in the standard library.
http://www.artima.com/weblogs/viewpost.jsp?thread=98196