Make perspectiveMode vary rendered size based on distance (and render states themselves on boards if sufficiently close also), use bitwise NOT and conditionalFlip function, from which use revelation to simplify axialDisparity, add two as-yet unused functions
I don't know what I was thinking in the axialDisparity function, in retrospect it looks like the kind of function I'd have made in Scratch where ternary operators didn't exist and I didn't want to use multiple lines (though in this situation, I suppose I only didn't want to reiterate them). Upon applying conditionalFlip, however, its nature was laid bare and I realised that it was wrong (it took a lot of thinking to verify that than to have the much cooler realisation about the axis==2 case (which I will put in a comment on this commit)).
The as-yet unused functions may perhaps also be reducible to a closed form by this method of studying their tree (this time ternary ones).
The space mode is currently set to the 8*8 KvK endgame, my framerate is low for it (due to my computer being slow and old (like me) and the O(n*(n-1)/2) repulsion computation (perhaps I will implement the fast multipole method)) but you can see a grid-like structure emerge and an overall shape form, even with eightfold symmetry reduction.
I will also find which direction it is from the initial spaceship position towards the diagram so you don't have to rotate around to look for it (if you couldn't tell, mapping quaternions to spatial coordinates has been done so far by manually adding inversions and reversals until it worked (to account for which there now is probably some change to the rotateVector matrix's indices and polarities, if you have the time)).