Add pawns (including cool promotion combination multiset algorithm)
I originally discovered a much less elegant form of the algorithm in my Spaceflight Simulator engine calculator a year ago, when I wanted to enumerate combinations of up to a maximum number of engines but didn't want duplicates (ie. [0,1] and [1,0]), and realised it can be done by counting in base b (number of engine types) except where, when a digit is incremented, all those following it become the same value as its new one, meaning the last one has a sum of b possibilities when the second-last is 0 or absent, then b-1 when it's 1, and so on, for the second-last to have a sum of b+(b-1)+...+1, ∑_(n=0)^(b-1) (b-n) =1/2*b*(b+1) when the third-last is 0 or absent, then (b-1)+(b-2)+...+1, and so on, for ∑_(n=0)^(b-1) ((b-n)*(n+1)) =1/6*b*(b+1)*(b+2) combinations, as you would expect a variable third digit causes it to become 1/24*b*(b+1)*(b+2)*(b+3), the closed form is, for n digits, ∏_(m=b)^(b+n) (m) /∏_(m=0)^(n) (m)=(b-1+n)!/((b-1)!*n!), or (b-1+n) choose (n), instead of b**n. When summing with respect to n (due to generating all combinations of fewer also), as I did in the engine calculator, the closed form is instead ((n+1)*(b+n)!-b!*(n+1)!)/(b*(b-1)!*(n+1)!).

In pawn move generation, I had the same problem for combinations of promotions, and used the same method of counting ascending runs (except descending due to little-endianness being more concise here) is used for each colour and each combination of promotions is found, then the usual subset function runs on that to account for captures.

Promotions currently work if you're God or in the CLI mode (with standard syntax with the = symbol) or viewing the state transition diagram, but not yet the GUI playing mode, you will have to be content with it crashing due to invalid dict key upon attempting to do so for now.

findPieceMoves now has a built-in option to return moves as tuples, instead of the stateMoves and stateTransitions list comprehension doing so itself, because they can now be of length 3 to contain the destination piece type also for promotions.