Realise eightfold cellular automaton state generation algorithm doesn't work, clean it up slightly, add RLE exporting and preliminary (also not working) chess symmetry function rewrite
The RLE function doesn't yet support encoding multiple $'s with the number of consecutive instances before a single one like it does for o's and b's.
The symmetry function rewrite is like the other tree structure functions, but with a regular if statement on the outermost layer instead of a ternary operator, due to a necessary allocation depending. Something is wrong (when it's used, KRvK on 8*8 is recognised as having maximum DTM 5 instead of 32), I thought about it much and haven't realised why yet.
Usually my commits add effectual things, this one only optimises a function that doesn't work and adds an optimisation that doesn't work for another, sorry.