Add preliminary support for toruses and modes for viewing nodes with domain colouring of DTM and king position
Toroidal boards have reduction by vertex transitivity now (as well as eightfold symmetry), 9 states instead of 464 in 8*8 KvK, great speedup
Currently, it doesn't do reduction in any axes if the board is an even width and the black king is on the square half the board width away from it in both axes, because it can't reduce in all but one line of symmetry and use an axial disparity as a tie breaker like usually, you must solve for all three simultaneously with something like the cellular automaton state generation layers algorithm. There is a canonical origin square (that of the white king), so it's easier than transitivity-reduced cellular automaton state generation (which can't be solved with the layers algorithm), at least.
Also, in state transition diagram mode, there are two new colour modes, one of winningness (like the previous one, but this time with brightness based on the DTM (weighted by how close it is (assuming perfect play)), and easier on the eyes), the other of the colour on the colour wheel corresponding with the direction from the centre to the white king's position, or from halfway between the bottom-left corner and the black king's if reducing by vertex-transitivity (because the white king is fixed and the black is in the bottom-left quadrant's bottom-right diagonal half).
You can also now use the shift key for stopping camera motion, the z key for pausing simulation, and the x key for viewing connections at half-length (making it clearer when one node is protruding from the rest of the graph whether it's itself a checkmate or just a losing one-legal-move state, because the green moves to it (from winning states) won't be obstructing the red one from it, or only if you've paused physics and want to see in which directions loops would be spinning).
The ternary operator tree form of the symmetry function is now fixed, the part where the white king is in the centre now has a copy with some changes for toruses, it looks like an inconcise duplicate now but will work for Klein bottles and real projective planes also, eventually (justifying it, I hope).
This commit also brings the line count to a nice, round number.