Add preliminary support for cellular automata (and state transition diagrams of which)
Conway's game of life is one of 2**18 outer-totalistic (OT) rules (cellular automata which each are defined by 9 booleans, each representing whether a cell with one of 9 sums of its eight neighbours (a king move away) becomes on if it's off, and another 9 for whether one with one of 9 sums remains on if it's already on), isotropic non-totalistic (INT) rules generalise this to 2**102, with births and survivals ('transitions') being dependent on the neighbour configuration (under eightfold symmetry) as well as sum. This supports them, and trivially also MAP rules (to which it converts OT and INT rules for simulation).
It works with all of the topological manifolds the tablebase one does. In a future commit, cycles will be enumerated (for still lifes and oscillators), and bounded boards will have all neighbouring cells outside the board check whether they turn on for each state and, if so, disqualify the state from the cycle enumeration function.

I found the OEIS sequence with closed form equations for the number of states there should be for an n*n bounded board under eightfold symmetry (or as they put it, "action of dihedral group of the square D_4"), which has shown my layers algorithm to be broken for 4*4 and larger boards (with oblique layers), I'm not sure whether it can be fixed by adjusting precedences, so run with eightfold symmetry reduction disabled for now.
This commit took a while because in a previous one, when optimising the positionReflect functions, I had erroneously believed Python supported BEDMAS, it actually reads multiplicative operations (and some others) left-to-right (as in C), which produces the same result, except for in floordiv (which I had used extensively, thus breaking most of them (and taking a long time to trace back to the reflect functions themselves and not the layers algorithm as I had initially assumed (due to having found many problems in))). https://docs.python.org/3/reference/expressions.html#operator-precedence

I don't expect to find anything new with this (especially not with Catagolue now existing), but the state transition diagram mode is interesting, I think (they look like ethereal space creatures unfurling). My inspiration (as mentioned in the readme) had been https://demonstrations.wolfram.com/CellularAutomatonStateTransitionDiagrams/, that does this purpose for the 1-dimensional Wolfram rule 150 (though I made the original Scratch chess piece vision program earlier), I don't know whether Stephen Wolfram's graph immersion algorithm is explained anywhere but it would provide a faster way than inverse-square law and Hooke's law simulation.

Six months ago, when I didn't know much about Python (and thus wrote inelegant programs), I had made a primitive version of this idea in https://github.com/DroneBetter/WolframTapeExhauster/ for my goal of finding a p13 oscillator (to which I was beaten by a human who found patterns and period-dividing mechanisms manually) in a bounded tape in the Wolfram rule 150, for which I had found analogue emulators in one such isotropic non-totalistic rule (that this program can now emulate).