Change to instead be for royal knights game
The game takes place on a 3*3 board (and is thus easily searchable), kings (beginning on a1 and a3 respectively) move as knights, bishops and rooks (on (b1 and b3) and (c1 and c3) respectively) can be placed upon the board (consuming a turn) by the player from whom they were captured (unlike crazyhouse, where they're transferred to the capturer's hand). Use CLI mode, standard crazyhouse syntax (placing pieces with ie. "B@b2") applies, it is a mate in 5 for white