Add still life/oscillator enumeration
States that would grow out of the board if it weren't bounded are disqualified from the search, but replacing the part in brackets in iterateCellular from line 690 to 692 with False will make it find oscillators specific to bounded boards (perhaps of interest, there is then a p3 oscillator in 4*4 in Conway's game of life, 4b$4o$obbo$boob!).
It reduces by displacements and reflections, also, so doesn't report any duplicates (now only four p2 oscillators (the blinker, toad, clock and beacon) in 4*4 CGoL boards instead of thirteen, and only one instance of the p16 in b3s34).
I'm thinking, because about half of all states will exceed the board's boundaries, removing them and all of their predecessor chains will make unreduced 3*3 and 4*4 boards' state transition diagrams feasible to simulate. The next commit will also have some more node colouring options for when states aren't rendered (including perhaps a very interesting one) instead of making them all grey.
Also, when working with unreduced states (which are conversely nicer because they can be indexed by summing the place values of cells (in technically O(lg(n)) time but too fast to care) and also generated by bitshifts and bitwise ANDings with 1 for each element, making the states list redundant for them), I think the diagram generation and oscillator enumeration could perhaps be done in a single pass (generating the state transitions in-place, perhaps not recording them at all if no state transition diagram is to be viewed) instead of multiple, so if I were to make it report results in real time (ie. for 5*5 boards, for which it would currently take months on my computer so would be uninteresting to be running for a long time without any reports, even if it were to be faster with future optimisations).