Add shut-the-box mode (with probabilistic tablebase and state transition diagrams)
In Shut the Box, you have a box with a lid comprised of a set of flaps numbered from 1 to 9, initially all open, each turn someone rolls a die (or two if a flap with a number exceeding 6 remains open) and closes any combination of open flaps with numbers summing to the dice sum, and it continues until all flaps have been closed and the box is shut. This is for the specific variant where the goal is to be the one to shut the box, not to accumulate the fewest points.
You would imagine it would be a simple task, but to play optimally from a given state requires knowledge of the winningness of all states reachable by making moves to choose the best one at each branch, similarly to chess, however winningnesses are not booleans assuming perfect play from both sides, but numbers between 0 and 1, due to its probabilistic nature. However, they can be found as closed fractions from recursively solving simple arithmetic for each state with all of its descendants' winningnesses known, accounting for the probability that no move is possible and it becomes the opponent's turn with the same winningness (inverted for you). Initially only the closed state is known, so only state-dice combinations with single open flaps are solvable (you would imagine also those with multiple open flaps, all simultaneously closable in a single move, but they only provide information for a single dice sum, not all, so can't be regressed from), then all with two open, and so on.
With 9 flaps and assuming play, in the variant where you don't change turns until you cannot make a legal move, there is a 23737591359877398294692797328494109466082417874622514538722483/46023439431450849098621618688813466452959545352645769768837120 chance that the first player wins, and in the other where you always alternate turns, there is a 65164008049389068064389926863232721272900040252761773269/130343949131039176368264671312660774887431339964953600000 chance. I believe I'm the first to discover this (however, Bill Butler found optimal strategies for the singleplayer variant by a similar method, I think, https://www.durangobill.com/ShutTheBox.html).
You can also view the state transition diagrams, with domain colouring of the dice sums (the flaps' colours don't go all the way around the colour wheel unless the number is a multiple of 6), it is also a hypercube, where a connection between states represents a decrement in axes with indices summing to its colour. (This property is only true if faces is at least flaps*(flaps+1)/2.)