Add bitwise cellular automaton simulation mode
Though I had made https://gist.github.com/DroneBetter/4f5d775e7c37f062ce750d4a8fefc84a (a modification of https://gist.github.com/DavidBuchanan314/acae2aab38953759aacc114b417ed0b9 that added support for INT rules and manifolds to be compatible with this program) already, this commit took a long time (much longer than expected) because there are still many things that had to be added, in which there were pieces of what should be simple arithmetic but that contained a great many things that can go wrong. This commit is provided as is with no guarantee of elegance, that will come later, it is only made because I got it working and want to be able to track all changes from here in case the optimisations break anything.
Many things have been optimised already, however, you must only check whether the edgemask (the union of the scrollmasks in which no scrolling is to occur and there is no gutter preservation) is nonzero, reducing the expression greatly, and for whatever reason the oscillator-finding part (that converts "loops" (state transition diagram cycles) to oscillators, finding their true period by simulating them in an un-symmetry-reduced plane) doesn't need to account for the last element being a copy of the first, I had been alternating the bitwise mode periodically (as it were :-) to check that everything was still working and found multiple times that it was either detecting periods as one iteration higher or lower, it took me a long time to realise that it was due to such a strange reason when the logic is the same.
It is probably slower with the bitwise mode enabled due to all non-SIMD conversion operations (though the insertion of bit "padding" can be done in logarithmic operations, being analogous to bit reversal), but some things involving recursive simulation will be faster, if and when I optimise them.
I suppose you could still simulate it with knight neighbourhoods (from precomputedKnightMoves) instead with bitwise enabled, but then you wouldn't be able to do the adding/ORing of the horizontally-added/ORed masks vertically, you would have to use more operations and scrollmasks would have to be two cells wide instead of one, which I don't care to create.