Add interface for chess mode (not finished)
The approximation of the number of states shown as you are placing pieces could be done trivially by finding the factorial of the board squares divided by the product of those of the numbers of occurrences of the different piece types, for each combination reachable from the current one, however a much better upper bound can be found by accounting for king nonattackingness and symmetry. I didn't want to compute these by explicitly finding them by running generateKingPositions only to take its output's length, and it seemed like the kind of thing that should be on OEIS. I had a look and couldn't find any, so added mine (for eightfold and twofold symmetry respectively, A357723 and A357740), they have more closed forms and recurrence relations there (though my inversion of the square matrices (from the quartic with a quadratic difference) to find them was redundant, the relation is the same as for all other quartics with quadratic differences).
It is extremely lacklustre currently, but has graphs of DTM showing the current position within them, and lets you use at least the chess mode in its entirety without pressing any keys.