Add eightfold reducer and Cburnett pieces
It occurred to me when I was making the cellular automaton mode that the eightfold reducer would be useful to others who wanted to make their own programs for searching for oscillators (or even simulating from all nonequivalent initial states in infinite planes with RLEs in apgsearch or something) in isotropic rules, but the comparable performance of doing so this way (as opposed to only iterating over all unreduced ones and discarding those that are not the minimal orientation of themselves) wouldn't be important enough to warrant it. Indeed, making such a complicated system as I had seemed something of a mistake. For 5*5 boards, generating all 4211744 nonequivalent states takes 56s with it on my computer, whereas all 33554432 unreduced ones take 460s (though this is with the trivial tuple comprehension in generateCellular()'s outermost else, I think an accumulate()-based one that recorded a single structure it mutated in-place could be faster)
However, upon learning of the Pólya enumeration theorem, I realised that, say on an odd-width board, by beginning with the innermost layer (a single cell) and using the theorem to compute the number of states with it turned off, you can deduce, given an index, whether this cell is on or off based on whether it's in the first or second half, and more generally which of the indexes of states obtainable by changing only the innermost cell it exceeds. This is all very simple, but after determining this, you can then search each "legal" state of the first orthogonally-displaced layer (ie. each that is the canonical representation of itself under symmetry), and determine how many nonequivalent states of the board excluding them and the innermost cell there are, under action of the symmetry group of reflections that it still allows. By only slightly changing the symmetry function, this could be trivially implemented, allowing any state to be indexed (and vice versa, by enacting the same process the other way around) by a series of arithmetic operations asymptotic to the number of layers (and thus cells). Being that bisections could obtain it in logarithmic time with respect to the length of the states list, which itself takes exponential space with respect to the cells, this isn't an improvement, and indeed, running __getitem__() on each index takes far longer than generateCellular(), which only takes logarithmic time with respect to cells for the mutation (which I will eventually implement as __next__() for slicing purposes), however this allows __getitem__() to run in the same order of time without need for generating and storing it all (taking, for 5*5, at the very least 100.4MiB of RAM (it would information-theoretically be 12.55MiB but the convention for processors is to store a byte per boolean), but much more due to it being Python and them not being arrays). I haven't seen this idea of mine implemented already (though I have seen some Catagolue censuses with soup-counts in A054247, so others must have considered this, at least), and it may have applications to multithreading (being that you could set each thread about simulating contiguous slices of even size) and even distributed computing.
Also, I have thought some more and realised that the application of the idea similar to that of tablebases, of generating and storing the entire state transition diagram and searching for cycles all simultaneously, is perhaps less efficient than considering each individually, being that it takes about logarithmic space to store all oscillators with respect to the time spent searching, and logarithmically many tiimes that to store numbers of occurrences of each (if you would like to), so perhaps I ought to make an option not to (especially considering this new tool).
This is only a preliminary commit, it is very inelegant (ie. the symmetry(), __getitem__() and __index__() functions contain a great deal of repetition (always repeat)) and will be reduced and optimised a great deal (as well as having some of the testing parts cleaned up), which will be fun (expanding and inlining things in order to simplify them)
Also, I'm very sorry for not having been so active here as of late, I have been busy with real life (and will continue to be so for the foreseeable future)

Regarding the images, now others using the tablebase program besides me will be able to see the nice lichess images instead of geometric shapes (but you don't get to use your imagination so much), though I don't know how to import svg's in Pygame at target resolutions instead of original ones (it seems you must load in and change the file metadata, I think), though it doesn't matter too much because chess isn't a very detail-intensive game and when there are hundreds or even thousands of instances being reprinted each frame in the state transition diagram mode, it's probably for the best.